Refactoring: 1.Push Down Method: Pushed sendBlock method of Sendable class to only the classes that are using sendBlock method
