+++ src/main/java/net/sacredlabyrinth/phaed/simpleclans/commands/data/Kills.java
@@ -3,6 +3,7 @@ package net.sacredlabyrinth.phaed.simpleclans.commands.data;
 import net.sacredlabyrinth.phaed.simpleclans.ChatBlock;
 import net.sacredlabyrinth.phaed.simpleclans.Helper;
 import net.sacredlabyrinth.phaed.simpleclans.SimpleClans;
+import net.sacredlabyrinth.phaed.simpleclans.managers.SettingsManager;
 import org.bukkit.ChatColor;
 import org.bukkit.entity.Player;
 import org.bukkit.scheduler.BukkitRunnable;
@@ -11,8 +12,7 @@ import org.jetbrains.annotations.NotNull;
 import java.util.Map;
 
 import static net.sacredlabyrinth.phaed.simpleclans.SimpleClans.lang;
-import static net.sacredlabyrinth.phaed.simpleclans.managers.SettingsManager.ConfigField.PAGE_CLAN_NAME_COLOR;
-import static net.sacredlabyrinth.phaed.simpleclans.managers.SettingsManager.ConfigField.PAGE_SEPARATOR;
+import static net.sacredlabyrinth.phaed.simpleclans.managers.SettingsManager.ConfigField.*;
 
 public class Kills extends Sendable {
 
@@ -59,5 +59,18 @@ public class Kills extends Sendable {
                 + " " + lang("kills", player) + " " + headColor +
                 Helper.generatePageSeparator(sm.getString(PAGE_SEPARATOR)));
         ChatBlock.sendBlank(player);
+    }
+
+    protected void sendBlock() {
+        SettingsManager sm = plugin.getSettingsManager();
+        boolean more = chatBlock.sendBlock(sender, sm.getInt(PAGE_SIZE));
+
+        if (more) {
+            plugin.getStorageManager().addChatBlock(sender, chatBlock);
+            ChatBlock.sendBlank(sender);
+            ChatBlock.sendMessage(sender, sm.getColored(PAGE_HEADINGS_COLOR) + lang("view.next.page", sender,
+                    sm.getString(COMMANDS_MORE)));
+        }
+        ChatBlock.sendBlank(sender);
     }
 }
