+++ src/main/java/net/sacredlabyrinth/phaed/simpleclans/commands/data/Sendable.java
@@ -30,18 +30,5 @@ public abstract class Sendable {
         subColor = sm.getColored(PAGE_SUBTITLE_COLOR);
     }
 
-    protected void sendBlock() {
-        SettingsManager sm = plugin.getSettingsManager();
-        boolean more = chatBlock.sendBlock(sender, sm.getInt(PAGE_SIZE));
-
-        if (more) {
-            plugin.getStorageManager().addChatBlock(sender, chatBlock);
-            ChatBlock.sendBlank(sender);
-            ChatBlock.sendMessage(sender, sm.getColored(PAGE_HEADINGS_COLOR) + lang("view.next.page", sender,
-                    sm.getString(COMMANDS_MORE)));
-        }
-        ChatBlock.sendBlank(sender);
-    }
-
     public abstract void send();
 }
